+++ src/main/java/net/helix/hlx/SignedFiles.java
@@ -1,10 +1,6 @@
 package net.helix.hlx;
 
-import net.helix.hlx.crypto.Sha3;
-import net.helix.hlx.crypto.Winternitz;
-import net.helix.hlx.crypto.Merkle;
-import net.helix.hlx.crypto.Sponge;
-import net.helix.hlx.crypto.SpongeFactory;
+import net.helix.hlx.crypto.*;
 import net.helix.hlx.model.Hash;
 import org.apache.commons.lang3.ArrayUtils;
 import org.bouncycastle.util.encoders.Hex;